requirements.txt
asgiref==3.3.1
Django==3.1.7
django-eav2==0.13.0
pytz==2021.1
sqlparse==0.4.1
